Fix: 복습문제 쿼리문 변경(#31)

+++ back/src/main/java/com/clover/recode/domain/statistics/service/StatisticsService.java
@@ -1,12 +1,11 @@
 package com.clover.recode.domain.statistics.service;
 
 import com.clover.recode.domain.statistics.dto.response.StatisticsListRes;
-import org.springframework.stereotype.Service;
 
 public interface StatisticsService {
 
     StatisticsListRes getStatisticsList(Long id);
 
-    int getReviewCnt(Long id);
+    Long getReviewCnt(Long id);
 
 }
